feat: MeshGen Blocky, mesh load, render init
stdx::cast<>()
VertexData::Clear()
World::GetCell()

+++ src/ethertia/world/Chunk.cpp
@@ -4,7 +4,7 @@
 
 
 
-Chunk::Chunk(World* world, glm::vec3 cp) : m_World(world), chunkpos(cp)
+Chunk::Chunk(World* world, glm::ivec3 cp) : m_World(world), chunkpos(cp)
 {
 
 }
